🔨refactor: page 컴포넌트를 서버 컴포넌트로 전환
